+++ packages/sdk/src/domain.ts
@@ -2,7 +2,14 @@ import type { Link } from "@imtbl/imx-sdk"
 import type { Address } from "@rarible/types"
 import type { BigNumberValue } from "@rarible/utils"
 import type { TransferRequest, TransferResponse } from "./nft/domain"
-import type { BuyRequest, BuyResponse, CancelOrderRequest, SellRequest } from "./order/domain"
+import type {
+	BuyRequest,
+	BuyResponse,
+	CancelOrderRequest,
+	CancelOrderResponse,
+	SellRequest,
+	SellResponse,
+} from "./order/domain"
 import type { BalanceRequestAssetType } from "./balance/balance"
 
 export type ImxEnv = "e2e" | "dev" | "staging" | "prod"
@@ -21,13 +28,13 @@ export type ImxFee = { recipient: string, percentage: number }
 export type ImxBlockchainTx = void
 
 export type ImxWallet = {
-	connect(): ReturnType<Link["setup"]>
+	registerImx(): ReturnType<Link["setup"]>
 }
 
 export type ImxOrderSdk = {
 	buy(request: BuyRequest): Promise<BuyResponse>
-	sell(request: SellRequest): Promise<ImxBlockchainTx>
-	cancel(request: CancelOrderRequest): Promise<ImxBlockchainTx>
+	sell(request: SellRequest): Promise<SellResponse>
+	cancel(request: CancelOrderRequest): Promise<CancelOrderResponse>
 }
 
 export type ImxNftSdk = {